third_party/WebKit/LayoutTests/external/wpt/html/dom/reflection-obsolete-expected.txt

This is a testharness.js-based test.
Found 2809 tests; 2385 PASS, 424 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS applet.title: 32 tests
PASS applet.lang: 32 tests
PASS applet.dir: 62 tests
PASS applet.className (<applet class>): 32 tests
PASS applet.hidden: 33 tests
PASS applet.accessKey: 32 tests
PASS applet.tabIndex: 24 tests
FAIL applet.align: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.align: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.align: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.align: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.align: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.align: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.align: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.align: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.align: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.align: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.align: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.align: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.align: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.align: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.align: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.align: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.align: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.align: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.align: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.align: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.align: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.align: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.align: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.align: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.align: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.align: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.align: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.align: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.align: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.align: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.align: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.align: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.alt: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.alt: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.alt: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.alt: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.alt: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.alt: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.alt: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.alt: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.alt: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.alt: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.alt: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.alt: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.alt: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.alt: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.alt: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.alt: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.alt: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.alt: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.alt: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.alt: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.alt: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.alt: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.alt: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.alt: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.alt: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.alt: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.alt: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.alt: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.alt: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.alt: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.alt: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.alt: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.archive: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.archive: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.archive: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.archive: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.archive: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.archive: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.archive: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.archive: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.archive: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.archive: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.archive: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.archive: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.archive: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.archive: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.archive: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.archive: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.archive: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.archive: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.archive: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.archive: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.archive: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.archive: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.archive: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.archive: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.archive: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.archive: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.archive: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.archive: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.archive: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.archive: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.archive: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.archive: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.code: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.code: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.code: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.code: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.code: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.code: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.code: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.code: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.code: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.code: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.code: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.code: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.code: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.code: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.code: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.code: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.code: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.code: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.code: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.code: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.code: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.code: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.code: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.code: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.code: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.code: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.code: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.code: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.code: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.code: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.code: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.code: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.codeBase: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.codeBase: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to "" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to " foo " assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/foo" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to "http://site.example/" assert_equals: IDL get expected (string) "http://site.example/" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to "//site.example/path???@#l" assert_equals: IDL get expected (string) "http://site.example/path???@#l" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to undefined assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/undefined" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to 7 assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/7" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to 1.5 assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/1.5" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to true assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/true" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to false assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/false" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/[object%20Object]" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to NaN assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/NaN" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to Infinity assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/Infinity" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to -Infinity assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/-Infinity" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to "\0" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to null assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/null" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-toString" but got (undefined) undefined
FAIL applet.codeBase: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-valueOf" but got (undefined) undefined
FAIL applet.codeBase: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.codeBase: IDL set to " foo " assert_equals: getAttribute() expected " foo " but got "test-valueOf"
FAIL applet.codeBase: IDL set to "http://site.example/" assert_equals: getAttribute() expected "http://site.example/" but got "test-valueOf"
FAIL applet.codeBase: IDL set to "//site.example/path???@#l" assert_equals: getAttribute() expected "//site.example/path???@#l" but got "test-valueOf"
FAIL applet.codeBase: IDL set to "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " assert_equals: getAttribute() expected "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " but got "test-valueOf"
FAIL applet.codeBase: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.codeBase: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.codeBase: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.codeBase: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.codeBase: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.codeBase: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.codeBase: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.codeBase: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.codeBase: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.codeBase: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.codeBase: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.codeBase: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.codeBase: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-valueOf" but got (object) object "test-valueOf"
FAIL applet.height: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.height: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.height: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.height: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.height: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.height: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.height: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.height: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.height: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.height: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.height: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.height: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.height: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.height: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.height: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.height: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.height: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.height: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.height: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.height: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.height: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.height: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.height: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.height: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.height: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.height: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.height: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.height: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.height: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.height: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.height: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.height: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.hspace: typeof IDL attribute assert_equals: expected "number" but got "undefined"
FAIL applet.hspace: IDL get with DOM attribute unset assert_equals: expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to -2147483649 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to -2147483648 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to -36 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to -1 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 0 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 1 assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 257 assert_equals: IDL get expected (number) 257 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 2147483647 assert_equals: IDL get expected (number) 2147483647 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 2147483648 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 4294967295 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 4294967296 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "-1" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "-0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "1" assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\t7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\v7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\f7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "﻿7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\n7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\r7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "᠎7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "　7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to undefined assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to 1.5 assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to true assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to false assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to object "[object Object]" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to NaN assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to Infinity assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to -Infinity assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to "\0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to object "2" assert_equals: IDL get expected (number) 2 but got (undefined) undefined
FAIL applet.hspace: setAttribute() to object "3" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.hspace: IDL set to 0 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.hspace: IDL set to 1 assert_equals: getAttribute() expected "1" but got "[object Object]"
FAIL applet.hspace: IDL set to 257 assert_equals: getAttribute() expected "257" but got "[object Object]"
FAIL applet.hspace: IDL set to 2147483647 assert_equals: getAttribute() expected "2147483647" but got "[object Object]"
FAIL applet.hspace: IDL set to "-0" assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.hspace: IDL set to 2147483648 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.hspace: IDL set to 4294967295 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.name: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.name: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.name: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.name: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.name: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.name: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.name: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.name: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.name: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.name: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.name: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.name: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.name: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.name: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.name: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.name: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.name: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.name: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.name: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.name: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.name: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.name: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.name: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.name: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.name: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.name: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.name: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.name: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.name: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.name: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.name: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.name: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
FAIL applet.object: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.object: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.object: setAttribute() to "" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.object: setAttribute() to " foo " assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/foo" but got (undefined) undefined
FAIL applet.object: setAttribute() to "http://site.example/" assert_equals: IDL get expected (string) "http://site.example/" but got (undefined) undefined
FAIL applet.object: setAttribute() to "//site.example/path???@#l" assert_equals: IDL get expected (string) "http://site.example/path???@#l" but got (undefined) undefined
FAIL applet.object: setAttribute() to "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.object: setAttribute() to undefined assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/undefined" but got (undefined) undefined
FAIL applet.object: setAttribute() to 7 assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/7" but got (undefined) undefined
FAIL applet.object: setAttribute() to 1.5 assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/1.5" but got (undefined) undefined
FAIL applet.object: setAttribute() to true assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/true" but got (undefined) undefined
FAIL applet.object: setAttribute() to false assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/false" but got (undefined) undefined
FAIL applet.object: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/[object%20Object]" but got (undefined) undefined
FAIL applet.object: setAttribute() to NaN assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/NaN" but got (undefined) undefined
FAIL applet.object: setAttribute() to Infinity assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/Infinity" but got (undefined) undefined
FAIL applet.object: setAttribute() to -Infinity assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/-Infinity" but got (undefined) undefined
FAIL applet.object: setAttribute() to "\0" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/reflection-obsolete.html" but got (undefined) undefined
FAIL applet.object: setAttribute() to null assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/null" but got (undefined) undefined
FAIL applet.object: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-toString" but got (undefined) undefined
FAIL applet.object: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-valueOf" but got (undefined) undefined
FAIL applet.object: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.object: IDL set to " foo " assert_equals: getAttribute() expected " foo " but got "test-valueOf"
FAIL applet.object: IDL set to "http://site.example/" assert_equals: getAttribute() expected "http://site.example/" but got "test-valueOf"
FAIL applet.object: IDL set to "//site.example/path???@#l" assert_equals: getAttribute() expected "//site.example/path???@#l" but got "test-valueOf"
FAIL applet.object: IDL set to "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " assert_equals: getAttribute() expected "\0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f " but got "test-valueOf"
FAIL applet.object: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.object: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.object: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.object: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.object: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.object: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.object: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.object: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.object: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.object: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.object: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.object: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.object: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "http://web-platform.test:8001/html/dom/test-valueOf" but got (object) object "test-valueOf"
FAIL applet.vspace: typeof IDL attribute assert_equals: expected "number" but got "undefined"
FAIL applet.vspace: IDL get with DOM attribute unset assert_equals: expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to -2147483649 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to -2147483648 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to -36 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to -1 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 0 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 1 assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 257 assert_equals: IDL get expected (number) 257 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 2147483647 assert_equals: IDL get expected (number) 2147483647 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 2147483648 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 4294967295 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 4294967296 assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "-1" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "-0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "1" assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\t7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\v7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\f7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "﻿7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\n7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\r7" assert_equals: IDL get expected (number) 7 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "᠎7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " 7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "　7" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to undefined assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to 1.5 assert_equals: IDL get expected (number) 1 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to true assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to false assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to object "[object Object]" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to NaN assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to Infinity assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to -Infinity assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to "\0" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to object "2" assert_equals: IDL get expected (number) 2 but got (undefined) undefined
FAIL applet.vspace: setAttribute() to object "3" assert_equals: IDL get expected (number) 0 but got (undefined) undefined
FAIL applet.vspace: IDL set to 0 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.vspace: IDL set to 1 assert_equals: getAttribute() expected "1" but got "[object Object]"
FAIL applet.vspace: IDL set to 257 assert_equals: getAttribute() expected "257" but got "[object Object]"
FAIL applet.vspace: IDL set to 2147483647 assert_equals: getAttribute() expected "2147483647" but got "[object Object]"
FAIL applet.vspace: IDL set to "-0" assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.vspace: IDL set to 2147483648 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.vspace: IDL set to 4294967295 assert_equals: getAttribute() expected "0" but got "[object Object]"
FAIL applet.width: typeof IDL attribute assert_equals: expected "string" but got "undefined"
FAIL applet.width: IDL get with DOM attribute unset assert_equals: expected (string) "" but got (undefined) undefined
FAIL applet.width: setAttribute() to "" assert_equals: IDL get expected (string) "" but got (undefined) undefined
FAIL applet.width: setAttribute() to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: IDL get expected (string) " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got (undefined) undefined
FAIL applet.width: setAttribute() to undefined assert_equals: IDL get expected (string) "undefined" but got (undefined) undefined
FAIL applet.width: setAttribute() to 7 assert_equals: IDL get expected (string) "7" but got (undefined) undefined
FAIL applet.width: setAttribute() to 1.5 assert_equals: IDL get expected (string) "1.5" but got (undefined) undefined
FAIL applet.width: setAttribute() to true assert_equals: IDL get expected (string) "true" but got (undefined) undefined
FAIL applet.width: setAttribute() to false assert_equals: IDL get expected (string) "false" but got (undefined) undefined
FAIL applet.width: setAttribute() to object "[object Object]" assert_equals: IDL get expected (string) "[object Object]" but got (undefined) undefined
FAIL applet.width: setAttribute() to NaN assert_equals: IDL get expected (string) "NaN" but got (undefined) undefined
FAIL applet.width: setAttribute() to Infinity assert_equals: IDL get expected (string) "Infinity" but got (undefined) undefined
FAIL applet.width: setAttribute() to -Infinity assert_equals: IDL get expected (string) "-Infinity" but got (undefined) undefined
FAIL applet.width: setAttribute() to "\0" assert_equals: IDL get expected (string) "\0" but got (undefined) undefined
FAIL applet.width: setAttribute() to null assert_equals: IDL get expected (string) "null" but got (undefined) undefined
FAIL applet.width: setAttribute() to object "test-toString" assert_equals: IDL get expected (string) "test-toString" but got (undefined) undefined
FAIL applet.width: setAttribute() to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (undefined) undefined
FAIL applet.width: IDL set to "" assert_equals: getAttribute() expected "" but got "test-valueOf"
FAIL applet.width: IDL set to " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " assert_equals: getAttribute() expected " \0\x01\x02\x03\x04\x05\x06\x07 \b\t\n\v\f\r\x0e\x0f \x10\x11\x12\x13\x14\x15\x16\x17 \x18\x19\x1a\x1b\x1c\x1d\x1e\x1f  foo " but got "test-valueOf"
FAIL applet.width: IDL set to undefined assert_equals: getAttribute() expected "undefined" but got "test-valueOf"
FAIL applet.width: IDL set to 7 assert_equals: getAttribute() expected "7" but got "test-valueOf"
FAIL applet.width: IDL set to 1.5 assert_equals: getAttribute() expected "1.5" but got "test-valueOf"
FAIL applet.width: IDL set to true assert_equals: getAttribute() expected "true" but got "test-valueOf"
FAIL applet.width: IDL set to false assert_equals: getAttribute() expected "false" but got "test-valueOf"
FAIL applet.width: IDL set to object "[object Object]" assert_equals: getAttribute() expected "[object Object]" but got "test-valueOf"
FAIL applet.width: IDL set to NaN assert_equals: getAttribute() expected "NaN" but got "test-valueOf"
FAIL applet.width: IDL set to Infinity assert_equals: getAttribute() expected "Infinity" but got "test-valueOf"
FAIL applet.width: IDL set to -Infinity assert_equals: getAttribute() expected "-Infinity" but got "test-valueOf"
FAIL applet.width: IDL set to "\0" assert_equals: getAttribute() expected "\0" but got "test-valueOf"
FAIL applet.width: IDL set to null assert_equals: getAttribute() expected "null" but got "test-valueOf"
FAIL applet.width: IDL set to object "test-toString" assert_equals: getAttribute() expected "test-toString" but got "test-valueOf"
FAIL applet.width: IDL set to object "test-valueOf" assert_equals: IDL get expected (string) "test-valueOf" but got (object) object "test-valueOf"
PASS marquee.title: 32 tests
PASS marquee.lang: 32 tests
PASS marquee.dir: 62 tests
PASS marquee.className (<marquee class>): 32 tests
PASS marquee.hidden: 33 tests
PASS marquee.accessKey: 32 tests
PASS marquee.tabIndex: 24 tests
PASS marquee.behavior: 32 tests
PASS marquee.bgColor: 32 tests
PASS marquee.direction: 32 tests
PASS marquee.height: 32 tests
PASS marquee.hspace: 62 tests
PASS marquee.scrollAmount: 62 tests
PASS marquee.scrollDelay: 62 tests
PASS marquee.trueSpeed: 33 tests
PASS marquee.vspace: 62 tests
PASS marquee.width: 32 tests
PASS frameset.title: 32 tests
PASS frameset.lang: 32 tests
PASS frameset.dir: 62 tests
PASS frameset.className (<frameset class>): 32 tests
PASS frameset.hidden: 33 tests
PASS frameset.accessKey: 32 tests
PASS frameset.tabIndex: 24 tests
PASS frameset.cols: 32 tests
PASS frameset.rows: 32 tests
PASS frame.title: 32 tests
PASS frame.lang: 32 tests
PASS frame.dir: 62 tests
PASS frame.className (<frame class>): 32 tests
PASS frame.hidden: 33 tests
PASS frame.accessKey: 32 tests
PASS frame.tabIndex: 24 tests
PASS frame.name: 32 tests
PASS frame.scrolling: 32 tests
PASS frame.src: 38 tests
PASS frame.frameBorder: 32 tests
PASS frame.longDesc: 38 tests
PASS frame.noResize: 33 tests
PASS frame.marginHeight: 32 tests
PASS frame.marginWidth: 32 tests
PASS dir.title: 32 tests
PASS dir.lang: 32 tests
PASS dir.dir: 62 tests
PASS dir.className (<dir class>): 32 tests
PASS dir.hidden: 33 tests
PASS dir.accessKey: 32 tests
PASS dir.tabIndex: 24 tests
PASS dir.compact: 33 tests
PASS font.title: 32 tests
PASS font.lang: 32 tests
PASS font.dir: 62 tests
PASS font.className (<font class>): 32 tests
PASS font.hidden: 33 tests
PASS font.accessKey: 32 tests
PASS font.tabIndex: 24 tests
PASS font.color: 32 tests
PASS font.face: 32 tests
PASS font.size: 32 tests
Harness: the test ran to completion.
